Broke out useSettersAsEventHandler to separate file
